+++ src/idl_gen_text.cpp
@@ -51,7 +51,7 @@ bool Print(T val, Type type, int /*indent*/, Type * /*union_type*/,
            const IDLOptions &opts, std::string *_text) {
   std::string &text = *_text;
   if (type.enum_def && opts.output_enum_identifiers) {
-    auto enum_val = type.enum_def->ReverseLookup(static_cast<int>(val));
+    auto enum_val = type.enum_def->ReverseLookup(static_cast<int64_t>(val));
     if (enum_val) {
       text += "\"";
       text += enum_val->name;
